+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Error messages now display the path of the file containing the problem.
+- Maps and modules with erroneous extra fields now have a custom error
+  message.
 
 ## v0.1.0 - 2019-04-15